Add a basic yaml config file for redis urls using the Vegas CLI.
The only catch is a chicken-and-egg with option parsing and app
construction.  We get around it with a shim app that accumulates
constructor arguments and lazily constructs the `Qless::ShardedUI`
instance.

The locked version of ruby-debug was also bumped allow it to work with
ruby 1.9.3.

I've also dropped a `before_run` hook that `load`ed the value of
`RESQUECONFIG` or the first positional argument if either was present.
Kick me and I'll make it come back.